Add wincrypt and common crypto backends
libyara supports three mutually exclusive crypto backends that it can use in its
hash and pe modules during compilation: OpenSSL, WinCrypt (windows)
or Common Crypto (macos).

The original configure.ac is in charge of finding the right crypto backend.

Change YARA_ENABLE_CRYPTO boolean flag to YARA_CRYPTO_LIB which specifies the
desired crypto backend. When unspecified, we default to the crypto backend of
the target os, which will likely always be available. The user can override
this default, for exemple to force the use of OpenSSL on a windows target that
has a Cygwin-like environement.

The hash module compilation now defaults to **Enable**, like in the original
libyara.